I added links for MAP and SubDyn in the FAST glue code. This involved adding lines in the FAST input file.
I updated the compile_FASt.bat file with the necessary changes (it runs); the makefile doens't run (partly because I need to figure out what to do with SubDyn's use of MKL, etc).

Note that this works with the AddedMass terms from HydroDyn (modified from Greg's branch); the accelerations do not work.

git-svn-id: https://windsvn2.nrel.gov/FAST/branches/BJonkman/Utilities/SimulationToolbox@475 10acb478-4768-415a-8850-bacdb5912d4d